pyproject.toml
[tool.poetry]
name        = "book-scraper-api"
version     = "0.1.0"
description = "API for querying books via scraping"
authors     = ["Everton <evertonvieira91@gmail.com>"]
license     = "MIT"
readme      = "README.md"
packages    = [{ include = "api" }]

[tool.poetry.dependencies]
python          = ">=3.12"
fastapi         = "0.116.1"
pydantic        = "2.11.7"
beautifulsoup4  = "4.12.3"
uvicorn = "^0.35.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "<4.0"
python-multipart = "^0.0.20"
sqlmodel = "^0.0.24"
pydantic-settings = "^2.10.1"
requests = "^2.32.4"
apscheduler = "^3.11.0"
lxml = "^6.0.0"
sklearn-utils = "^0.0.15"

[tool.poetry.dev-dependencies]
pytest = "*"
black  = "*"
flake8 = "*"

[build-system]
requires    = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
